move the desc_buffer to the commit class to avoid descructor invoke
